chore: rename `AccountLayout` to `SettingsLayout`
