+++ packages/webapp/components/layouts/SettingsLayout/Profile/VerifiedCompanyBadge/VerifiedCompanyBadgeSection.tsx
