+++ packages/webapp/pages/settings/feed/preferences.tsx
@@ -3,10 +3,10 @@ import type { ReactElement } from 'react';
 import type { NextSeoProps } from 'next-seo';
 
 import { FeedSettingsContentPreferencesSection } from '@dailydotdev/shared/src/components/feeds/FeedSettings/sections/FeedSettingsContentPreferencesSection';
-import { AccountPageContainer } from '../../../components/layouts/AccountLayout/AccountPageContainer';
+import { AccountPageContainer } from '../../../components/layouts/SettingsLayout/AccountPageContainer';
 import { defaultOpenGraph, defaultSeo } from '../../../next-seo';
 import { getTemplatedTitle } from '../../../components/layouts/utils';
-import { getFeedSettingsLayout } from '../../../components/layouts/AccountLayout/FeedSettingsLayout';
+import { getFeedSettingsLayout } from '../../../components/layouts/SettingsLayout/FeedSettingsLayout';
 
 const AccountManageSubscriptionPage = (): ReactElement => {
   return (
