+++ packages/webapp/pages/settings/invite.tsx
@@ -48,9 +48,9 @@ import { useLazyModal } from '@dailydotdev/shared/src/hooks/useLazyModal';
 import { LazyModal } from '@dailydotdev/shared/src/components/modals/common/types';
 import { useLogContext } from '@dailydotdev/shared/src/contexts/LogContext';
 import { PlusUser } from '@dailydotdev/shared/src/components/PlusUser';
-import AccountContentSection from '../../components/layouts/AccountLayout/AccountContentSection';
-import { AccountPageContainer } from '../../components/layouts/AccountLayout/AccountPageContainer';
-import { getAccountLayout } from '../../components/layouts/AccountLayout';
+import AccountContentSection from '../../components/layouts/SettingsLayout/AccountContentSection';
+import { AccountPageContainer } from '../../components/layouts/SettingsLayout/AccountPageContainer';
+import { getSettingsLayout } from '../../components/layouts/SettingsLayout';
 import { GiftIcon, InviteIcon } from '@dailydotdev/shared/src/components/icons';
 import { defaultSeo } from '../../next-seo';
 import { getTemplatedTitle } from '../../components/layouts/utils';
@@ -213,7 +213,7 @@ const AccountInvitePage = (): ReactElement => {
   );
 };
 
-AccountInvitePage.getLayout = getAccountLayout;
+AccountInvitePage.getLayout = getSettingsLayout;
 AccountInvitePage.layoutProps = { seo };
 
 export default AccountInvitePage;
